+++ src/pages/changelog/index.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import Layout from "@theme/Layout";
 import releases from "@site/src/data/releases.json";
-import { ReleaseBody } from "./markdown";
+import { ReleaseBody } from "./_markdown";
 import "./index.css";
 
 type Release = {
